feat: fairly stable implementation of EEPROM sync

+++ src/state/XYMapState.h
@@ -2,8 +2,8 @@
 #define XYMAPSTATE_H
 
 typedef struct XYMapState {
-    int16_t x;
-    int16_t y;
+    int x;
+    int y;
 } XYMapState;
 
 #endif // XYMAPSTATE_H
